ARM: dts: msm: add a refgen proxy enable request for Kona
Add proxy-consumer properties to the refgen regulator device node
so that it is kept enabled until all consumers have a chance to
issue requests for it.  This regulator is shared by camera and
display hardware.  It must remain enabled after exiting from
boot loaders in order for continuous splash screen to function.

Change-Id: I992719554a5c497e02512d13766467066e6e1de5