msm: camera: reqmgr: increase the rd idx if no lower pd device
For link with maximum pipeline delay of 1 e.g.,
TPG use case or sensors with pipeline delay of 1,
if the request is not submitted before 2
consecutive triggers we do not get chance to
increment rd idx, in the mean time the slot which
was last applied will be reset and we will not be
able to apply request even if new requests are scheduled.

This will cause the camera to not apply any request
further, hence increasing the rd idx if no lower pd
devices are pending will fix the issue.

CRs-Fixed: 2626743
Change-Id: Ie60ac2751585d9a8019d36d9b0124f2d36483224
Signed-off-by: Tejas Prajapati <tpraja@codeaurora.org>