ARM: dts: msm: Add dpdm regulator node for USB on Scuba
Add the dpdm regulator node for USB to ensure that the
glue driver can put the core in known low power mode
if the device is boot up with no cable connected.

Also, this helps to prevent the incorrect charger type
detection by the PHY driver on Scuba during bootup with
cable connected by preventing the glue driver interfering
with the PHY driver's charger detection.

Change-Id: I021c0b0e4c892ac3daccd2a01c14c36bd58a0ec3