lto: Allow LTO for sys_ni.c
At some earlier point LTO was disabled for sys_ni.c to handle the system
call aliases without problems. This is not needed anymore, now that we
express syscall aliaseing in C instead of assembler. In fact
current LTO builds require this file to be LTOed because of some strange
external symbol weak resolution quirks in gcc 10.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Change-Id: I21988524230cebd5bd37cd72a7463c0bc281d0ea
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>