qcacld-3.0: Retain the stainfo ref count in case of re-connect
Currently in SAP case if a station disconnects and re-connects
to the SAP, in the process of register sta, SAP clears
previous reference of the station and re-initializes the sta
info. Because of this when the thread which actually took the
reference, tries to release the reference, is not able to release
the reference as the ref count is already zet to zero as part
of the reconnect which results in undefined behaviour.

To address this issue, retain the ref count of the sta info
in case of re-connect from the same station.

Change-id: Ie9558ca8e59601b48f9be7fff37cc530874ab5ad
CRs-Fixed: 2730782