qcacld-3.0: Update rssi on connect big data logging
Currently driver does not update rssi on connect, driver
only updates rssi on get_station command. If driver does
not receive get_station command between connect and disconnect
in that case driver will send invalid rssi value to userspace
as part of the big data logging command.
For example, in current scenario if station is connected to an ap
and if this station moves away which will change the rssi and
finally it will result in disconnect because of low rssi, at this
time rssi value stored in driver will be the value corresponding
to the last connection(low rssi). Now if this station connects to
some other ap and this connection fails at EAPOL so for this
connection driver will not receive any get station command and
it will not update rssi corresponding to the second ap and when
driver receives big data stats command it will give incorrect rssi
value of previous ap (low rssi).

To address above issue update rssi in host driver at connect
time also.

Change-Id: Ieac4b42033e8ad011e36384b0cd52a2c1f0a352e
CRs-Fixed: 2798813