qcacld-3.0: Validate the duration and interval for P2P_SET_NOA
The following are the firmware expectations for the parameters
through p2p_set_noa:
If count > 1 , the duration should be smaller than interval.
If count = 1, the duration can be same or smaller as interval.
Hence, set the duration as interval if it is configured
greater than interval.
This commit adds the checks accordingly.
Fails the command if the above conditions do not pass.

Change-Id: Ic273b0167c3551cd4b21d4c6dc7f31ba5329bebc
CRs-Fixed: 2941170